Moving tests outside of tests.py
