+++ scheduler/tests.py
@@ -391,147 +391,6 @@ class CreateAccountTestCase(TestCase):
         self.assertFalse(response.context['user'].is_authenticated)
 
 
-class TestEditAccount(unittest.TestCase):
-    def setUp(self):
-        self.app = AdminAssignmentPage()
-        self.user = UserTable(firstName="John", lastName="Doe", email="email@gmail.com", phone="262-724-8212",
-                         address="some address", userType="instructor")
-        self.user.save()
-        self.userAccount = User(username="johnD", password="password123", email=self.user.email)
-        self.userAccount.save()
-
-    def tearDown(self):
-        self.user.delete()
-        self.userAccount.delete()
-
-    def test_editAccount(self):
-        newUser = self.app.editAccount(self.user.email, "newemail@uwm.com", "1234567890", "123 street", "TA")
-        self.assertEqual(newUser.email, "newemail@uwm.com")
-        self.assertEqual(newUser.phone, "1234567890")
-        self.assertEqual(newUser.address, "123 street")
-        self.assertEqual(newUser.userType, "TA")
-
-    def test_editMissingAccount(self):
-        self.assertRaises(ValueError, self.app.editAccount, "nonexistent@email.com", "newemail@uwm.com", "1234567890", "123 street", "TA")
-
-
-class TestDeleteAccount(TestCase):
-    def setUp(self):
-        self.app = AdminAssignmentPage()
-        # user 1
-        self.user1 = UserTable(firstName="John", lastName="Doe", email="John@gmail.com", phone="262-724-8212",
-                               address="some address", userType="Instructor")
-        self.user1.save()
-        self.user1Account = User(username="john", email=self.user1.email, password="password123")
-        self.user1Account.save()
-        # user 2
-        self.user2 = UserTable(firstName="Jeff", lastName="Doe", email="Jeff@gmail.com", phone="262-724-8212",
-                               address="some address", userType="Instructor")
-        self.user2.save()
-        self.user2Account = User(username="jeff", email=self.user2.email, password="password123")
-        self.user2Account.save()
-
-    def test_deleteAccount(self):
-        result = self.app.deleteAccount(self.user1Account.id, self.user1Account.id)
-        self.assertEqual("Account deleted successfully", result)
-
-    def test_deleteAccountInvalidAccount(self):
-        result = self.app.deleteAccount(99999999, 99999999)
-        self.assertEqual("Failed to delete account", result)
-
-    def test_deleteAccountEmptyArguments(self):
-        with self.assertRaises(ValueError):
-            self.app.deleteAccount("", "")
-
-    def test_deleteAccountEmptyUsername(self):
-        with self.assertRaises(ValueError):
-            self.app.deleteAccount("", self.user1.email)
-
-    def test_deleteAccountEmptyEmail(self):
-        with self.assertRaises(ValueError):
-            self.app.deleteAccount(self.user1Account.username, "")
-
-    def test_deleteAccountWrongEmail(self):
-        result = self.app.deleteAccount(self.user1Account.id, self.user2Account.id)
-        self.assertEqual("username/email match error", result)
-
-    def test_deleteTwoAccount(self):
-        result = self.app.deleteAccount(self.user1Account.id, self.user1Account.id)
-        result2 = self.app.deleteAccount(self.user2Account.id, self.user2Account.id)
-        self.assertEqual("Account deleted successfully", result, result2)
-
-    def test_invalidArg(self):
-        result = self.app.deleteAccount(1, "ID")
-        self.assertEqual("username/email match error", result)
-
-    def test_deleteSameAccountTwice(self):
-        result = self.app.deleteAccount(self.user1Account.id, self.user1Account.id)
-        result2 = self.app.deleteAccount(self.user1Account.id, self.user1Account.id)
-        self.assertEqual("Failed to delete account", result2)
-
-
-class TestDeleteAccountACCEPTANCE(TestCase):
-    def __init__(self, methodName: str = "runTest"):
-        super().__init__(methodName)
-        self.client = None
-
-    def setUp(self):
-        self.app = AdminAssignmentPage()
-        self.user1 = UserTable(firstName="adminTest", lastName="adminTest", email="adminTest@gmail.com",
-                               phone="adminTest",
-                               address="adminTest", userType="admin")
-        self.user1.save()
-        self.user1Account = User(username="adminTest", password="adpassword", email=self.user1.email)
-        self.user1Account.save()
-
-        self.user2 = UserTable(firstName="deleteTest", lastName="deleteTest", email="deleteTest@gmail.com",
-                               phone="deleteTest",
-                               address="deleteTest", userType="ta")
-        self.user2.save()
-        self.user2Account = User(username="deleteTest", password="delpassword", email=self.user2.email)
-        self.user2Account.save()
-
-    def tearDown(self):
-        # Clean up test data
-        self.user1.delete()
-        self.user1Account.delete()
-        self.user2.delete()
-        self.user2Account.delete()
-
-    def test_adminAccManagement_Admin(self):
-        request = RequestFactory().get(reverse('adminAccManagement'))  # Use reverse to get the URL
-        request.user = self.user1Account
-
-        response = AdminAccManagement.as_view()(request)
-
-        # Check if the response status code is 200
-        self.assertEqual(response.status_code, 200)
-
-    def test_adminAccManagement_Ta(self):
-        request = RequestFactory().get(reverse('adminAccManagement'))  # Use reverse to get the URL
-        request.user = self.user2Account
-
-        response = AdminAccManagement.as_view()(request)
-
-        # Check if the response status code Redirect
-        self.assertEqual(response.status_code, 302)
-
-    def test_adminAccManagement_DeleteAccount(self):
-        request = RequestFactory().post(reverse('adminAccManagement'))
-        request.user = self.user1Account
-
-        requestCopy = request.POST.copy()
-        requestCopy['deleteAccountName'] = self.user2.id
-        requestCopy['deleteAccountEmail'] = self.user2.id
-        requestCopy['deleteAccBtn'] = 'Delete'
-
-        request.POST = requestCopy
-
-        response = AdminAccManagement.as_view()(request)
-
-        self.assertContains(response, 'Failed to delete account')
-
-
 class TestDeleteCourse(unittest.TestCase):
     def setUp(self):
         self.admin_page = AdminAssignmentPage()
