+++ scheduler/tests/acceptance/test_deleteAccount.py
@@ -1,0 +1,71 @@
+import unittest
+from django.test import RequestFactory
+from django.contrib.auth.models import User
+
+from scheduler.views import AdminAccManagement
+from adminAssignmentPage import AdminAssignmentPage
+from scheduler.models import UserTable, CourseTable, UserCourseJoinTable, SectionTable
+from django.test import TestCase
+from django.urls import reverse
+
+
+class TestDeleteAccountACCEPTANCE(TestCase):
+    def __init__(self, methodName: str = "runTest"):
+        super().__init__(methodName)
+        self.client = None
+
+    def setUp(self):
+        self.app = AdminAssignmentPage()
+        self.user1 = UserTable(firstName="adminTest", lastName="adminTest", email="adminTest@gmail.com",
+                               phone="adminTest",
+                               address="adminTest", userType="admin")
+        self.user1.save()
+        self.user1Account = User(username="adminTest", password="adpassword", email=self.user1.email)
+        self.user1Account.save()
+
+        self.user2 = UserTable(firstName="deleteTest", lastName="deleteTest", email="deleteTest@gmail.com",
+                               phone="deleteTest",
+                               address="deleteTest", userType="ta")
+        self.user2.save()
+        self.user2Account = User(username="deleteTest", password="delpassword", email=self.user2.email)
+        self.user2Account.save()
+
+    def tearDown(self):
+        # Clean up test data
+        self.user1.delete()
+        self.user1Account.delete()
+        self.user2.delete()
+        self.user2Account.delete()
+
+    def test_adminAccManagement_Admin(self):
+        request = RequestFactory().get(reverse('adminAccManagement'))  # Use reverse to get the URL
+        request.user = self.user1Account
+
+        response = AdminAccManagement.as_view()(request)
+
+        # Check if the response status code is 200
+        self.assertEqual(response.status_code, 200)
+
+    def test_adminAccManagement_Ta(self):
+        request = RequestFactory().get(reverse('adminAccManagement'))  # Use reverse to get the URL
+        request.user = self.user2Account
+
+        response = AdminAccManagement.as_view()(request)
+
+        # Check if the response status code Redirect
+        self.assertEqual(response.status_code, 302)
+
+    def test_adminAccManagement_DeleteAccount(self):
+        request = RequestFactory().post(reverse('adminAccManagement'))
+        request.user = self.user1Account
+
+        requestCopy = request.POST.copy()
+        requestCopy['deleteAccountName'] = self.user2.id
+        requestCopy['deleteAccountEmail'] = self.user2.id
+        requestCopy['deleteAccBtn'] = 'Delete'
+
+        request.POST = requestCopy
+
+        response = AdminAccManagement.as_view()(request)
+
+        self.assertContains(response, 'Failed to delete account')
